Refactoring surface code, still broken
