+++ client/src/modules/view/components/surface/index.js
@@ -1,0 +1,2 @@
+export * from './region-surface';
+export * from './background-surface';
